+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Update bundler in the lock file in the Bundler manager.
 - Update development dependencies.
 
 ## 0.1.1 — May 24th, 2025